requirements.txt
torch
accelerate
huggingface_hub
gradio
transformers
spaces
docling-core
python-dotenv